Rename package to cppcmd
