+++ CMakeLists.txt
@@ -1,16 +1,16 @@
 cmake_minimum_required(VERSION 3.10)
-project(cppshell CXX)
+project(cppcmd CXX)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_BINARY_DIR})
 
-option(CPPSHELL_HAS_STREAM_OPERATORS "Enable usage of stream operators" ON)
+option(CPPCMD_HAS_STREAM_OPERATORS "Enable usage of stream operators" ON)
 
-add_library(cppshell INTERFACE)
+add_library(cppcmd INTERFACE)
 
-if (CPPSHELL_HAS_STREAM_OPERATORS)
-    target_compile_definitions(cppshell INTERFACE CPPSHELL_HAS_STREAM_OPERATORS)
+if (CPPCMD_HAS_STREAM_OPERATORS)
+    target_compile_definitions(cppcmd INTERFACE CPPCMD_HAS_STREAM_OPERATORS)
 endif()
-target_include_directories(cppshell INTERFACE include)
+target_include_directories(cppcmd INTERFACE include)
 
 if (BUILD_TESTING)
     enable_testing()
