+++ apps/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(TARGET_NAME ${PROJECT_NAME}-demo)
 add_executable(${TARGET_NAME} main.cpp)
-target_link_libraries(${TARGET_NAME} cppshell)
+target_link_libraries(${TARGET_NAME} cppcmd)
 
 target_compile_options(${TARGET_NAME} PRIVATE
   $<$<CXX_COMPILER_ID:MSVC>:/W4 /WX /EHsc>
