+++ conanfile.py
@@ -2,11 +2,11 @@ import os
 from conans import ConanFile, CMake, tools
 
 
-class CppShellConan(ConanFile):
-    name = "cppshell"
+class CppCmdConan(ConanFile):
+    name = "cppcmd"
     version = "0.1.0"
     settings = "os", "compiler", "arch", "build_type"
-    url = "https://github.com/remysalim/cppshell"
+    url = "https://github.com/remysalim/cppcmd"
     license = "MIT"
     description = "Simple cpp command interpreter header-only library"
     topics = ("header-only", "interpreter", "cpp")
