+++ test_package/example.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include <cppshell/CommandInterpreter.hpp>
+#include <cppcmd/CommandInterpreter.hpp>
 
 int main() {
     std::cout << "Hello world!" << std::endl;
